+++ opendevin/runtime/__init__.py
@@ -1,5 +1,3 @@
-from .docker.local_box import LocalBox
-from .docker.ssh_box import DockerSSHBox
 from .e2b.sandbox import E2BBox
 from .sandbox import Sandbox
